pyproject.toml
# DocDo — Systematic Review Pipeline
# ====================================
#
# Python library for running the PRISMA 2020 systematic review pipeline
# for "3D Organ Segmentation from CT Scans using Deep Learning".
#
# Installation (development mode):
#   pip install -e ".[dev]"
#
# Usage:
#   docdo fetch-pubmed
#   docdo deduplicate
#   docdo screen -i data.csv -o results.csv
#   docdo screen-batch -i data.csv
#   docdo compile
#   docdo stats

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docdo"
version = "0.1.0"
description = "Systematic review pipeline for 3D organ segmentation from CT using deep learning"
readme = "operations/README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "DocDo Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "click>=8.0",            # CLI framework
    "openai>=1.0",           # OpenAI API (screening & extraction)
    "python-dotenv>=1.0",    # .env file loading
    "pandas>=2.0",           # Data manipulation
    "numpy>=1.24",           # Numerical computation
    "requests>=2.28",        # HTTP requests (DOI verification)
    "biopython>=1.80",       # PubMed search via Entrez
    "PyMuPDF>=1.23",         # PDF text extraction (fitz)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
analysis = [
    "scipy>=1.10",
    "matplotlib>=3.0",
    "seaborn>=0.12",
]
full = [
    "httpx>=0.24",           # Async HTTP (optional)
    "arxiv>=2.0",            # arXiv API (optional)
]

[project.scripts]
docdo = "docdo.cli:main"

[project.urls]
Documentation = "documentation/"
Repository = "https://github.com/game-guild/docdo-paper-research"

[tool.hatch.build.targets.wheel]
packages = ["operations/src/docdo"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["operations/tests"]
addopts = "-v --cov=docdo"
